Fix ZUZ 5th and 6th issue - services in separate files, called from controller, not from models
